Made it deployable on a device.
A number of changes to make the app deployable on an actual device.
There were problems with MEF and RX libraries. The app worked fine in a
emulator by the deployment to a device was failing. MEF portable
libraries were not targeting WP8 - had to re-build it from the sources
and include the WP8 support. RX had a bug that was fixed in the latest
version - removed the RX from the Lib folder and referenced the latest
via NuGet.